+++ lib/guard/jasmine.rb
@@ -26,7 +26,7 @@ module Guard
         :server_timeout   => 15,
         :port             => nil,
         :rackup_config    => nil,
-        :jasmine_url      => 'http://localhost:8888/jasmine',
+        :jasmine_url      => nil,
         :timeout          => 10,
         :spec_dir         => 'spec/javascripts',
         :notification     => true,
@@ -69,7 +69,8 @@ module Guard
     # @option options [Hash] :run_all options overwrite options when run all specs
     #
     def initialize(watchers = [], options = { })
-      options[:jasmine_url] = "http://localhost:#{ options[:port] }/jasmine" if options[:port] && !options[:jasmine_url]
+      options[:port] ||= Jasmine.find_free_server_port
+      options[:jasmine_url] = "http://localhost:#{ options[:port] }/jasmine" unless options[:jasmine_url]
       options = DEFAULT_OPTIONS.merge(options)
       options[:specdoc] = :failure if ![:always, :never, :failure].include? options[:specdoc]
       options[:server] ||= :auto