+++ lib/guard/jasmine/cli.rb
@@ -31,7 +31,6 @@ module Guard
       method_option :port,
                     :type => :numeric,
                     :aliases => '-p',
-                    :default => 3001,
                     :desc => 'Server port to use'
 
       method_option :server_env,
@@ -64,7 +63,6 @@ module Guard
       method_option :url,
                     :type => :string,
                     :aliases => '-u',
-                    :default => 'http://localhost:3001/jasmine',
                     :desc => 'The url of the Jasmine test runner'
 
       method_option :timeout,
@@ -103,10 +101,10 @@ module Guard
         paths = [options.spec_dir] if paths.empty?
 
         runner_options = {}
-        runner_options[:jasmine_url] = options.url
+        runner_options[:port] = options.port || CLI.find_free_server_port
+        runner_options[:jasmine_url] = options.url || "http://localhost:#{ runner_options[:port] }/jasmine"
         runner_options[:phantomjs_bin] = options.bin || CLI.which('phantomjs')
         runner_options[:timeout] = options.timeout
-        runner_options[:port] = options.port
         runner_options[:server_env] = options.server_env
         runner_options[:server_timeout] = options.server_timeout
         runner_options[:spec_dir] = options.spec_dir