+++ spec/guard/jasmine/runner_spec.rb
@@ -7,7 +7,10 @@ describe Guard::Jasmine::Runner do
   let(:runner) { Guard::Jasmine::Runner }
   let(:formatter) { Guard::Jasmine::Formatter }
 
-  let(:defaults) { Guard::Jasmine::DEFAULT_OPTIONS.merge({ :phantomjs_bin => '/usr/local/bin/phantomjs' }) }
+  let(:defaults) { Guard::Jasmine::DEFAULT_OPTIONS.merge({
+      :jasmine_url => 'http://localhost:8888/jasmine',
+      :phantomjs_bin => '/usr/local/bin/phantomjs' })
+  }
 
   let(:phantomjs_error_response) do
     <<-JSON